refactoring notes
