+++ webview-ui/sid/src/notes/data/bpmJiffiex.json
@@ -1,0 +1,186 @@
+[
+  { "mm": 900, "w": 16, "h": 8, "q": 4, "e": 2, "s": 1, "t32": 0, "t64": 0 },
+  { "mm": 600, "w": 24, "h": 12, "q": 6, "e": 3, "s": 0, "t32": 0, "t64": 0 },
+  { "mm": 450, "w": 32, "h": 16, "q": 8, "e": 4, "s": 2, "t32": 1, "t64": 0 },
+  { "mm": 360, "w": 40, "h": 20, "q": 10, "e": 5, "s": 0, "t32": 0, "t64": 0 },
+  { "mm": 300, "w": 48, "h": 24, "q": 12, "e": 6, "s": 3, "t32": 0, "t64": 0 },
+  { "mm": 257, "w": 56, "h": 28, "q": 14, "e": 7, "s": 0, "t32": 0, "t64": 0 },
+  { "mm": 225, "w": 64, "h": 32, "q": 16, "e": 8, "s": 4, "t32": 2, "t64": 1 },
+  { "mm": 200, "w": 72, "h": 36, "q": 18, "e": 9, "s": 0, "t32": 0, "t64": 0 },
+  { "mm": 180, "w": 80, "h": 40, "q": 20, "e": 10, "s": 5, "t32": 0, "t64": 0 },
+  { "mm": 163, "w": 88, "h": 44, "q": 22, "e": 11, "s": 0, "t32": 0, "t64": 0 },
+  { "mm": 150, "w": 96, "h": 48, "q": 24, "e": 12, "s": 6, "t32": 3, "t64": 0 },
+  {
+    "mm": 138,
+    "w": 104,
+    "h": 52,
+    "q": 26,
+    "e": 13,
+    "s": 0,
+    "t32": 0,
+    "t64": 0
+  },
+  {
+    "mm": 128,
+    "w": 112,
+    "h": 56,
+    "q": 28,
+    "e": 14,
+    "s": 7,
+    "t32": 0,
+    "t64": 0
+  },
+  {
+    "mm": 120,
+    "w": 120,
+    "h": 60,
+    "q": 30,
+    "e": 15,
+    "s": 0,
+    "t32": 0,
+    "t64": 0
+  },
+  {
+    "mm": 112,
+    "w": 128,
+    "h": 64,
+    "q": 32,
+    "e": 16,
+    "s": 8,
+    "t32": 4,
+    "t64": 2
+  },
+  {
+    "mm": 105,
+    "w": 136,
+    "h": 68,
+    "q": 34,
+    "e": 17,
+    "s": 0,
+    "t32": 0,
+    "t64": 0
+  },
+  {
+    "mm": 100,
+    "w": 144,
+    "h": 72,
+    "q": 36,
+    "e": 18,
+    "s": 9,
+    "t32": 0,
+    "t64": 0
+  },
+  { "mm": 94, "w": 152, "h": 76, "q": 38, "e": 19, "s": 0, "t32": 0, "t64": 0 },
+  {
+    "mm": 90,
+    "w": 160,
+    "h": 80,
+    "q": 40,
+    "e": 20,
+    "s": 10,
+    "t32": 5,
+    "t64": 0
+  },
+  { "mm": 85, "w": 168, "h": 84, "q": 42, "e": 21, "s": 0, "t32": 0, "t64": 0 },
+  {
+    "mm": 81,
+    "w": 176,
+    "h": 88,
+    "q": 44,
+    "e": 22,
+    "s": 11,
+    "t32": 0,
+    "t64": 0
+  },
+  { "mm": 78, "w": 184, "h": 92, "q": 46, "e": 23, "s": 0, "t32": 0, "t64": 0 },
+  {
+    "mm": 75,
+    "w": 192,
+    "h": 96,
+    "q": 48,
+    "e": 24,
+    "s": 12,
+    "t32": 6,
+    "t64": 3
+  },
+  {
+    "mm": 72,
+    "w": 200,
+    "h": 100,
+    "q": 50,
+    "e": 25,
+    "s": 0,
+    "t32": 0,
+    "t64": 0
+  },
+  {
+    "mm": 69,
+    "w": 208,
+    "h": 104,
+    "q": 52,
+    "e": 26,
+    "s": 13,
+    "t32": 0,
+    "t64": 0
+  },
+  {
+    "mm": 66,
+    "w": 216,
+    "h": 108,
+    "q": 54,
+    "e": 27,
+    "s": 0,
+    "t32": 0,
+    "t64": 0
+  },
+  {
+    "mm": 64,
+    "w": 224,
+    "h": 112,
+    "q": 56,
+    "e": 28,
+    "s": 14,
+    "t32": 7,
+    "t64": 0
+  },
+  {
+    "mm": 62,
+    "w": 232,
+    "h": 116,
+    "q": 58,
+    "e": 29,
+    "s": 0,
+    "t32": 0,
+    "t64": 0
+  },
+  {
+    "mm": 60,
+    "w": 240,
+    "h": 120,
+    "q": 60,
+    "e": 30,
+    "s": 15,
+    "t32": 0,
+    "t64": 0
+  },
+  {
+    "mm": 58,
+    "w": 248,
+    "h": 124,
+    "q": 62,
+    "e": 31,
+    "s": 0,
+    "t32": 0,
+    "t64": 0
+  },
+  {
+    "mm": 56,
+    "w": 256,
+    "h": 128,
+    "q": 64,
+    "e": 32,
+    "s": 16,
+    "t32": 8,
+    "t64": 4
+  }
+]
